altered FileSystemObject, moved File and Directory classes inside FileSystemObject.py for simplicity
